--- docs/demo-py.rst
@@ -1,10 +1,0 @@
-:orphan:
-
-demo.py
-=======
-
-.. literalinclude:: _static/demo.py
-   :language: python
-   :linenos:
-
-:download:`Download source (demo.py) <_static/demo.py>`
