+++ docs/shallow-water-source.rst
@@ -1,0 +1,10 @@
+:orphan:
+
+:file:`examples/shallow_water.py`
+=================================
+
+.. literalinclude:: ../examples/shallow_water.py
+   :language: python
+   :linenos:
+
+:download:`Download source (shallow_water.py) <../examples/shallow_water.py>`
